chore: rename _shell to shell
