+++ core/components/fields/radio/index.tsx
@@ -1,7 +1,7 @@
 import type TField from "@/core/types/fields";
 import { useId } from "react";
 import Fields from "..";
-import Shell from "../_shell";
+import Shell from "../shell";
 
 type Props = TField<{
   value?: string;
